+++ src/main/java/com/wootube/ioi/service/util/FileUploader.java
@@ -5,54 +5,28 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.Optional;
 
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.CannedAccessControlList;
-import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.wootube.ioi.service.exception.FileUploadException;
 
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
-@Component
-public class FileUploader {
-	private static final String DIRECTORY_NAME = "wootube";
+public interface FileUploader {
+	String uploadCloud(File uploadFile);
 
-	@Qualifier(value = "amazonS3Client")
-	private final AmazonS3 amazonS3Client;
+	void deleteFile(String fileName);
 
-	public FileUploader(AmazonS3 amazonS3Client) {
-		this.amazonS3Client = amazonS3Client;
-	}
-
-	@Value("${cloud.aws.s3.bucket}")
-	private String bucket;
-
-	public String uploadFile(MultipartFile multipartFile) {
+	default String uploadFile(MultipartFile multipartFile) {
 		try {
 			File uploadFile = convert(multipartFile)
 					.orElseThrow(FileUploadException::new);
-			return upload(uploadFile, DIRECTORY_NAME);
+			String videoUrl = uploadCloud(uploadFile);
+			uploadFile.delete();
+			return videoUrl;
 		} catch (IOException e) {
 			throw new FileUploadException();
 		}
 	}
 
-	private String upload(File uploadFile, String directoryName) {
-		String fileName = directoryName + "/" + uploadFile.getName();
-		String videoUrl = uploadCloud(uploadFile, fileName);
-		uploadFile.delete();
-		return videoUrl;
-	}
-
-	private String uploadCloud(File uploadFile, String fileName) {
-		amazonS3Client.putObject(new PutObjectRequest(bucket, fileName, uploadFile)
-				.withCannedAcl(CannedAccessControlList.PublicRead));
-		return amazonS3Client.getUrl(bucket, fileName).toString();
-	}
-
-	private Optional<File> convert(MultipartFile file) throws IOException {
+	default Optional<File> convert(MultipartFile file) throws IOException {
 		File convertFile = new File(file.getOriginalFilename());
 		if (convertFile.createNewFile()) {
 			try (FileOutputStream fos = new FileOutputStream(convertFile)) {
@@ -61,9 +35,5 @@ public class FileUploader {
 			return Optional.of(convertFile);
 		}
 		return Optional.empty();
-	}
-
-	public void deleteFile(String originFileName) {
-		amazonS3Client.deleteObject(bucket, DIRECTORY_NAME + "/" + originFileName);
 	}
 }
